Add files via upload
Who: Society as a whole


What: A decline in embracing those who make our everyday lives better and witnessing the joy sounds bring upon others in our community

Where: In the inner cities


Why: It's important to discuss so that we can have memorable photos of those who gives us the soundtracks to our lives. To also witness the happiness in others around the world when resonating with sounds and vibrations.